requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
django-filter>=23.2
Pillow>=10.0
python-decouple>=3.8
pytz
sqlparse
tzdata
